requirements.txt
mysql-connector-python>=8.0.26
pandas>=1.3.0
numpy>=1.21.0
tushare>=1.2.89
pyinstaller>=5.6.2 